Add styling on each option element in the harvester list
since some browsers use the same style as used on the select element, others only
display a white background. This will color the entire dropdownlist either red or green,
according to the selected harvester - which is misleading.

Now the dropdownlist uses red or green text for disabled or enabled harvesters (but only
for browsers supporting styling of the option element.)

Auditors: jbn, atm