Handle mixed submitter summaries due to aborted summary retrievals
This happens when you type in a new submitter while the previous one is still loding summaries.
In that case, the final list will contain data for both the previous and the current submitter..

Sometimes, the count for the previous would be incorrect due to an aborted query, but not always..

This fix applies self-healing to the list so that the list of summaries is groomed during inserts.

Auditors: jbn, atm